requirements.txt
Flask==2.3.2
httplib2==0.20.2
Jinja2==3.1.2
oauthlib==3.2.0
PyJWT==2.3.0
Werkzeug==2.3.6
flask
flask-sqlalchemy
flask-cors
